pyproject.toml
[project]
name = "plane-to-jira"
version = "0.1.0"
description = "Migrate projects from Plane to JIRA, preserving issue IDs, comments, media, and assignees"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "markdownify>=1.2.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[tool.uv]
package = true

[project.scripts]
plane-to-jira = "plane_to_jira.cli:main"
